// clients/web/src/features/layout/composables/controller/appLayoutControllerTypes.ts
export type AppLayoutControllerActions = {
  message: {
    goToMessage: (channelId: string, messageId: string) => void;
  };
  navigation: {
    openDm: (userId: string) => void | Promise<void>;
  };
};